requirements.txt
Django==1.10.2
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-bootstrap3-datetimepicker==2.3
django-sendfile==0.3.11
django-user-accounts==1.3.1
jsonfield==1.0.3
pinax-eventlog==1.1.2
pinax-theme-bootstrap==7.11.0
pinax-webanalytics==2.0.4
pytz==2016.7
